Extract wireMarkdownActions and improve attach-markdown.sh comments
Code quality review findings:
- Extract duplicated markdown event delegation from home.ts and view.ts
  into wireMarkdownActions() in markdown-actions.ts
- Remove unused imports (handleMarkdownDownload, handleMarkdownCopy,
  logError) from view.ts and home.ts
- Add pre-flight check description to attach-markdown.sh header comment
- Document COLLECTION_PATH production-only intent
- Clarify STORAGE_PATH vs GCS_DEST relationship
